moved MergeAction into Merge namespace, changed SqlSyntax.FindObjectId to void from int
