+++ PostulateV1/Core.csproj
@@ -52,7 +52,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Abstract\MergeAction.cs" />
+    <Compile Include="Merge\MergeAction.cs" />
     <Compile Include="Abstract\SeedData.cs" />
     <Compile Include="Abstract\SqlDb_ChangeTracking.cs" />
     <Compile Include="Abstract\SqlDb_Copy.cs" />
